Add ZFS Health section
